+++ blocks/src/affiliate-grid/js/edit.js
@@ -5,14 +5,10 @@ export default function Edit( props ) {
 
 	const { className } = props;
 
-	if ( className === undefined ) {
-		className = 'wp-block';
-	}
-
 	return (
 		<Fragment>
 			<div
-				className={ `${ className } affiliate-grid row row-cols-1 row-cols-md-2 affiliate-items` }
+				className={ `${ className } wp-block affiliate-grid row row-cols-1 row-cols-md-2 affiliate-items` }
 			>
 				<InnerBlocks
 					template={ [
