+++ blocks/src/spoiler/css/editor.scss
@@ -1,4 +1,5 @@
-.wp-block-lez-library-spoilers {
+.wp-block-lez-library-spoilers,
+.lwtv-spoilers {
 	color: #721c24;
 	background-color: #f8d7da;
 	border-color: #f5c6cb;
