+++ blocks/src/spoiler/css/style.scss
